requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
backports.tarfile==1.2.0
black==25.1.0
build==1.2.2.post1
CacheControl==0.14.2
cachetools==5.5.2
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
cleo==2.1.0
click==8.1.8
colorama==0.4.6
crashtest==0.4.1
distlib==0.3.9
Django==5.2
django-debug-toolbar==5.1.0
django-rest-framework==0.1.0
django-wait-for-db==1.0.6
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
dulwich==0.22.8
fastjsonschema==2.21.1
filelock==3.18.0
findpython==0.6.3
flake8==5.0.4
flake8-quotes==3.3.1
flake8-variables-names==0.0.5
frozenlist==1.5.0
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
inflection==0.5.1
iniconfig==2.1.0
installer==0.7.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keyring==25.6.0
mccabe==0.7.0
mixins==0.1.4
more-itertools==10.7.0
msgpack==1.1.0
multidict==6.2.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
pbs-installer==2025.4.9
pep8-naming==0.13.2
pkginfo==1.12.1.2
platformdirs==4.3.7
pluggy==1.5.0
poetry==2.1.2
poetry-core==2.1.2
propcache==0.3.1
psycopg==3.2.6
psycopg-binary==3.2.6
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycodestyle==2.9.1
pyflakes==2.5.0
PyJWT==2.9.0
pyproject-api==1.9.0
pyproject_hooks==1.2.0
pytest==8.3.5
pytest-django==4.11.1
pytest-docker==3.2.1
python-dotenv==1.1.0
pytz==2025.2
pywin32-ctypes==0.2.3
PyYAML==6.0.2
RapidFuzz==3.13.0
referencing==0.36.2
requests==2.32.3
requests-toolbelt==1.0.0
rpds-py==0.24.0
shellingham==1.5.4
sniffio==1.3.1
sqlparse==0.5.3
tomlkit==0.13.2
tox==4.25.0
trove-classifiers==2025.4.11.15
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
virtualenv==20.30.0
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
